Made lots of enhancements to opening notes.
Made it so opening a new note from within the note editor brings up a confermation prompt to save the note first before opening another one. This prevents accidental loss of work and ensures intentional switching of unopened notes.
Replaced the old Open File Dialog with a new and improved Menu-Based list of notes. The app now scans data/notes for all txt files and displays them in an easy to view list. This allows for more control over how notes are selected and opened in the main window, and the note editor.